# packages/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Hoopoe Daemon API
  version: 0.1.0
  description: |
    Hoopoe daemon ↔ desktop contract per `plan.md §2.6`.

    Conventions (declared once here, applied across endpoints):

    - **Errors** use RFC 7807 `application/problem+json` (`#/components/schemas/Problem`).
    - **Idempotency** for retryable writes: clients pass `Idempotency-Key` (a stable
      ULID/UUID); the daemon dedupes within `IdempotencyKey.windowSeconds` and replays
      the original response on retry. Required on every non-GET that mutates state.
    - **Schema versions** every persisted entity declares `schemaVersion`. The daemon
      runs migrations at startup with backup + rollback (`plan.md §10.3`).
    - **Capability gating** every UI feature is gated on capability IDs reported by
      `/v1/capabilities` (§2.8). A capability that parses but cannot satisfy a
      contract test is `degraded`, not `ok`.
    - **WS events** the WebSocket envelope is documented as JSON Schema in
      `events/ws-envelope.schema.json` (sidecar; OpenAPI 3.1 cannot model the
      event-stream shape adequately).
    - **Source of truth** this spec is authoritative; both the TS client
      (`packages/schemas/ts/`) and the Go types (`packages/schemas/go/`) are
      generated from it. Hand-maintained duplicate shapes in `apps/desktop` or
      `apps/daemon` are forbidden (`plan.md §3 'Shared'`).
  license:
    name: UNLICENSED
servers:
  - url: http://127.0.0.1:0
...
  - url: http://127.0.0.1:8765
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List projects this VPS hosts.
      parameters:
        - name: cursor
          in: query
          required: false
          schema: { $ref: '#/components/schemas/Cursor' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Project list (paginated).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectListResponse' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Projects]
      operationId: createProject
      summary: Import an existing VPS checkout into Hoopoe.
      description: |
        Registers a project rooted at an existing VPS Git checkout. Hoopoe v1
        requires an origin remote, initializes `.hoopoe/`, runs `br init` when
        `.beads/` is missing, records detected project tools, and returns the
        persisted project registry row.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreateRequest' }
      responses:
        '201':
          description: Project imported.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get a single project by ID.
      responses:
        '200':
          description: Project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/projects/{projectId}/readiness:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    get:
      tags: [Projects]
      operationId: getProjectReadiness
      summary: Gate-readiness state — which §4.2 invariants pass / fail.
      description: |
        Used by the stage chrome to gate Plan → Beads → Swarm → Hardening
        transitions. Each gate reports its individual checks so Diagnostics can
        explain "what's missing" without a separate round-trip.
      responses:
        '200':
          description: Readiness snapshot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectReadiness' }
        default: { $ref: '#/components/responses/Problem' }

  # ─────────────────────────────────────────────────────────────────────────
  # Plans (§7.1) — read; create/lock land in subsequent commits.
  # ─────────────────────────────────────────────────────────────────────────
  /v1/projects/{projectId}/plans:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    get:
      tags: [Plans]
      operationId: listPlans
      summary: List plans for a project (locked + draft).
      parameters:
        - name: cursor
          in: query
          required: false
          schema: { $ref: '#/components/schemas/Cursor' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Plan list (paginated).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanListResponse' }
        default: { $ref: '#/components/responses/Problem' }

  # ─────────────────────────────────────────────────────────────────────────
  # Beads — read via `br --json` adapter; mutations via /v1/beads/{id}.
  # ─────────────────────────────────────────────────────────────────────────
  /v1/projects/{projectId}/beads:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    get:
      tags: [Beads]
      operationId: listBeads
      summary: List beads for a project.
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by bead status (multi-value supported).
          explode: true
          schema:
            type: array
            items: { $ref: '#/components/schemas/BeadStatus' }
        - name: cursor
          in: query
          required: false
          schema: { $ref: '#/components/schemas/Cursor' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: Bead list (paginated).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BeadListResponse' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/projects/{projectId}/beads/{beadId}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
      - name: beadId
        in: path
        required: true
        description: Beads ID (e.g., `hp-r3i`).
        schema: { type: string, pattern: '^[a-z0-9-]+$' }
    get:
      tags: [Beads]
      operationId: getBead
      summary: Get a single bead, with edges.
      responses:
        '200':
          description: Bead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bead' }
        default: { $ref: '#/components/responses/Problem' }

  # ─────────────────────────────────────────────────────────────────────────
  # Jobs (§2.7) — long-running daemon work registry.
  # ─────────────────────────────────────────────────────────────────────────
...
  /v1/jobs/{jobId}/cancel:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string }
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Jobs]
      operationId: cancelJob
      summary: Cancel a job (SIGTERM → grace → SIGKILL on the process group).
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCancelRequest' }
      responses:
        '202':
          description: Cancel acknowledged. Final state arrives via the job event channel.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        default: { $ref: '#/components/responses/Problem' }

  # ─────────────────────────────────────────────────────────────────────────
  # Approvals (§5.3) — Hoopoe-policy + ingested DCG verdicts in one queue.
  # ─────────────────────────────────────────────────────────────────────────
  /v1/projects/{projectId}/approvals:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
    get:
      tags: [Approvals]
      operationId: listApprovals
      summary: List pending + recently-decided approvals for a project.
      parameters:
        - name: state
          in: query
          required: false
          description: Filter by approval state.
          explode: true
          schema:
            type: array
            items: { $ref: '#/components/schemas/ApprovalState' }
        - name: cursor
          in: query
          required: false
          schema: { $ref: '#/components/schemas/Cursor' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Approval list (paginated).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApprovalListResponse' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/projects/{projectId}/approvals/{approvalId}/approve:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
      - name: approvalId
        in: path
        required: true
        schema: { type: string }
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Approvals]
      operationId: approveApproval
      summary: Approve a pending approval.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalDecisionRequest' }
      responses:
        '200':
          description: Approval decided.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        default: { $ref: '#/components/responses/Problem' }

  # ─────────────────────────────────────────────────────────────────────────
  # Events (§2.6) — replay stream for the WS reconnect path.
  # The bidirectional WebSocket itself is not modeled here (OpenAPI 3.x
  # cannot describe it); see `events/ws-envelope.schema.json` for the
  # standalone envelope JSON Schema that mirrors `components.WsEventEnvelope`.
  # ─────────────────────────────────────────────────────────────────────────
...
  /v1/projects/{projectId}/tending/actionplans:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Tending]
      operationId: submitActionPlan
      summary: Submit a typed ActionPlan from a tending agent (§8.3.1).
      description: |
        The daemon validates `actions[].kind` against the closed action set
        in `tending-actions.yaml`, evaluates policy + approval rules, and
        either accepts (returning a Job) or rejects with a `problem+json`
        whose `code` indicates the failure (e.g.,
        `tending.action_kind_unknown`, `approval.required`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionPlan' }
      responses:
        '202':
          description: ActionPlan accepted; Job tracks execution.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        default: { $ref: '#/components/responses/Problem' }

  # ─────────────────────────────────────────────────────────────────────────
  # Provider plugins (§6.2, §13) — abstract VPS provisioning catalog.
  # Daemon HTTP wrappers around the in-process `schemas.ProviderPlugin`
  # Go interface (registered via `init()` in plugin packages). v1 ships
  # with Contabo only (hp-9fo).
  # ─────────────────────────────────────────────────────────────────────────
...
  /v1/projects/{projectId}/approvals/{approvalId}/deny:
    parameters:
      - $ref: '#/components/parameters/ProjectIdPath'
      - name: approvalId
        in: path
        required: true
        schema: { type: string }
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Approvals]
      operationId: denyApproval
      summary: Deny a pending approval.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalDecisionRequest' }
      responses:
        '200':
          description: Approval decided.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        default: { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HMAC-signed session token (`plan.md §5.2`)
      description: |
        30-day bearer minted by exchanging a single-use pairing token at
        `/v1/auth/bootstrap/bearer`. Stored on the desktop via Electron
        `safeStorage.encryptString`. WebSocket connects use a 5-min `wsToken`
        from `/v1/auth/ws-token` instead.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Stable client-generated key (ULID/UUID) for safe retries. The daemon
        dedupes by key within a sliding window (default 24h) and replays the
        original status + body. Required on retryable writes; clients that omit
        it on a write that turns out to be retryable will receive a
        `precondition-failed` problem on the second attempt.
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      description: Project identifier (slug or ULID).
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9_\-]+$'
  responses:
    Problem:
      description: RFC 7807 problem+json error envelope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Cursor:
      type: string
      description: |
        Opaque pagination cursor. Use the `next` cursor from a list response
        verbatim; do not parse it. Cursors are stable for ≥24h.
...
    ProjectCreateRequest:
      type: object
      description: |
        Request body for importing an existing project checkout into the daemon
        registry. `Idempotency-Key` is supplied as an HTTP header, not in this
        JSON body.
      required: [rootPath]
      properties:
        id:
          type: string
          description: 'Optional stable project ID override; omitted values are daemon-generated.'
        name:
          type: string
          description: 'Optional display name; defaults to the project root basename.'
        slug:
          type: string
          description: 'Optional route-safe slug; defaults to a slugified name.'
        rootPath:
          type: string
          description: 'Absolute path to an existing VPS Git checkout.'
        vpsId:
          type: string
          description: 'Optional VPS host ID; defaults to the local daemon host.'
        desktopMirrorPath:
          type: string
          description: 'Read-only desktop sync mirror path once the desktop clone is created.'
        allowNoLanguageManifest:
          type: boolean
          description: 'Permit documentation-only repos without package manifests during readiness checks.'
    Project:
      type: object
      description: |
        One project on this VPS. Lifecycle states + gate invariants are
        canonical (§4); the daemon never lets a stage flip until the gate
        invariants pass.
      required: [id, slug, name, vpsId, repo, lifecycleState, schemaVersion]
      properties:
        schemaVersion: { $ref: '#/components/schemas/SchemaVersion' }
        id: { type: string }
        slug:
          type: string
          description: 'URL-safe project slug (used in routes + Activity panel).'
        name: { type: string, description: 'Human-readable project name.' }
        vpsId: { type: string, description: 'ID of the VpsHost hosting this project.' }
        repo: { $ref: '#/components/schemas/ProjectRepoRef' }
        lifecycleState: { $ref: '#/components/schemas/ProjectLifecycleState' }
        agentsManifestPresent:
          type: boolean
          description: 'Whether AGENTS.md was found on import.'
        hoopoeInitialized:
          type: boolean
          description: 'Whether the project has a populated `.hoopoe/` directory.'
        toolDetectionDone: { type: boolean }
        importedAt: { type: string, format: date-time }
        lastActivityAt: { type: string, format: date-time }
    ProjectListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Project' }
        page: { $ref: '#/components/schemas/PageMeta' }
...
    ProjectReadiness:
      type: object
      description: |
        Composite readiness snapshot per §4.2. `gates` is keyed in canonical
        order; the desktop renders them as a checklist with the first failing
        gate highlighted.
      required: [projectId, schemaVersion, gates, checkedAt]
      properties:
        schemaVersion: { $ref: '#/components/schemas/SchemaVersion' }
        projectId: { type: string }
        currentLifecycleState: { $ref: '#/components/schemas/ProjectLifecycleState' }
        gates:
          type: array
          items: { $ref: '#/components/schemas/ProjectReadinessGate' }
        checkedAt: { type: string, format: date-time }

    # ─────────────────────────────────────────────────────────────────────────
    # Plans (§7.1)
    # ─────────────────────────────────────────────────────────────────────────
...
    PlanListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Plan' }
        page: { $ref: '#/components/schemas/PageMeta' }

    # ─────────────────────────────────────────────────────────────────────────
    # Beads (read via `br --robot-*` adapter)
    # ─────────────────────────────────────────────────────────────────────────
    BeadStatus:
      type: string
      enum: [open, in_progress, blocked, closed, deleted]
    BeadIssueType:
      type: string
      enum: [task, bug, feature, epic, question, docs]
...
    Bead:
      type: object
      description: |
        One bead from `br`. Edges are cached snapshots; the canonical edge
        truth lives in `.beads/issues.jsonl`.
      required: [id, title, status, priority, issueType, schemaVersion]
      properties:
        schemaVersion: { $ref: '#/components/schemas/SchemaVersion' }
        id: { type: string, description: 'Beads ID, e.g., `hp-r3i`.' }
        title: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/BeadStatus' }
        priority: { $ref: '#/components/schemas/BeadPriority' }
        issueType: { $ref: '#/components/schemas/BeadIssueType' }
        sourceRepo: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { type: string }
        dependencies:
          type: array
          items: { $ref: '#/components/schemas/BeadDependencyEdge' }
        dependents:
          type: array
          items: { $ref: '#/components/schemas/BeadDependencyEdge' }
    BeadListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Bead' }
        page: { $ref: '#/components/schemas/PageMeta' }
...
    JobCancelRequest:
      type: object
      properties:
        reason: { type: string, description: 'Optional human-readable note.' }
        graceSeconds:
          type: integer
          minimum: 0
          maximum: 60
          default: 5
          description: 'SIGTERM → SIGKILL escalation window.'

    # ─────────────────────────────────────────────────────────────────────────
    # CommandSpec + Approval (§5.3)
    # ─────────────────────────────────────────────────────────────────────────
...
    ApprovalState:
      type: string
      enum: [pending, approved, denied, expired, revoked]
...